requirements.txt
# Core dependencies - keeping it minimal
pandas>=1.5.0,<2.0.0
numpy>=1.23.0,<2.0.0
python-dotenv>=0.20.0
snowflake-connector-python>=3.0.0
loguru>=0.7.0